+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 0.4.2
-- Update workspace activation event to require a `sorbet/config` file specifically (behavior equivalance).   
+- Update workspace activation event to require a `sorbet/config` file specifically (behavior equivalance). 
+- Improve **Create package file** quickfix for Sorbet 3705.
 
 ## 0.4.1
 - Upgrade minimum VS Code version to 1.125.